requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
chromadb==0.3.29
click==8.1.7
clickhouse-connect==0.7.19
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
duckdb==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.85.1
filelock==3.16.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.0
h11==0.14.0
hnswlib==0.8.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.24.7
humanfriendly==10.0
idna==3.10
importlib_metadata==7.2.1
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langchain-openai==0.0.5
langsmith==0.0.92
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.8.1
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
onnxruntime==1.19.2
openai==1.45.1
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.3
posthog==3.6.6
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
pulsar-client==3.5.0
pure_eval==0.2.3
pyarrow==17.0.0
pydantic==1.10.18
pydeck==0.9.1
Pygments==2.18.0
pyreadline3==3.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data==0.6.3
starlette==0.20.4
streamlit==1.30.0
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.5.2
tokenizers==0.19.1
toml==0.10.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.30.6
validators==0.34.0
watchdog==5.0.2
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0.1
yarl==1.11.1
zipp==3.20.2
zstandard==0.23.0
